+++ api/src/llms/interactions/chatInteraction.ts
@@ -1,6 +1,6 @@
 import LLMInteraction from 'api/llms/baseInteraction.ts';
 import type LLM from '../providers/baseLLM.ts';
-import { AnthropicModel } from 'api/types.ts';
+//import type { AnthropicModel } from 'api/types.ts';
 import type { LLMSpeakWithOptions, LLMSpeakWithResponse } from 'api/types.ts';
 import type { ConversationId } from 'shared/types.ts';
 import type { AuxiliaryChatContent } from 'api/logEntries/types.ts';
@@ -46,6 +46,8 @@ class LLMChatInteraction extends LLMInteraction {
 
 		//this.conversationLogger.logAuxiliaryMessage(messageId, prompt);
 
+		//speakOptions = { model: this.projectConfig.defaultModels!.chat, ...speakOptions };
+		if (!this.model) this.model = this.projectConfig.defaultModels!.chat;
 		logger.debug(`ChatInteraction: chat - calling llm.speakWithRetry for ${messageId}`);
 		const response = await this.llm.speakWithRetry(this, speakOptions);