+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- support for OpenAI provider
+- support for DeepSeek provider
 
 
 ## [0.6.0] - 2025-01-08